started gd voxelshape basics

+++ Library/Source/Library/Voxel/BinaryVolume.h
@@ -8,7 +8,7 @@ namespace Library
     class BinaryVolume
     {
       public:
-        glm::ivec3        dimension;
+        glm::u64vec3      dimension;
         glm::dvec3        origin;
         glm::dvec3        size;
         std::vector<bool> data;
